style: swiftlint 설정 변경, 파일 분리
