+++ OwnMyWay/OwnMyWay/Model/Record.swift
@@ -1,0 +1,17 @@
+//
+//  Record.swift
+//  OwnMyWay
+//
+//  Created by 김우재 on 2021/11/04.
+//
+
+import Foundation
+
+struct Record {
+    var uuid: UUID?
+    var content: String?
+    var date: Date?
+    var latitude: Double?
+    var longitude: Double?
+    var photoURL: URL?
+}
